requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
astroid==2.4.2
autopep8==1.5.4
boto3==1.18.15
botocore==1.21.15
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.2.5
django-bootstrap-form==3.4
django-cleanup==5.2.0
django-crispy-forms==1.10.0
django-debug-toolbar==2.2
django-filter==2.3.0
django-progressive-web-app==0.1.1


djangorestframework==3.12.4
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
mysql-connector-python==8.0.22
Pillow==8.0.1
protobuf==3.14.0
psycopg2==2.9.1
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.4
reportlab==3.5.67
requests==2.25.1
s3transfer==0.5.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
twilio==6.58.0
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1

